More robustness additions.
When updating we now check to see if index.json has changed. If it hasn't
we then check for changes to resources (basically the sprites), and update
those! So now we can get a sort of sub-refresh, i.e. the images are
refreshed but the objects aren't, so any dynamic objects won't lose their
state.