Introducing my own error stream.
Piping to stderr is all well and good, but we can do better! CernVM-Graphics
now with a Stream doubler so you can pipe to stderrgfx.txt whilst also
storing any errors in memory - very useful if you want to display the errors
and not just put them in a file.

(Note, changes to the Makefiles had to be made, so that mac one is now
untested).