More graceful degredation and debug stream
I killed off most of the uses of Errors::fatal as the program closing is not
a user friendly way of doing things. What's more the errors are now a little
more verbose (also commenting on what they're course of action is), and
there is a debug stream as well as an error stream.